requirements.txt
Pillow==10.4.0
requests==2.32.3
numpy==2.1.1